water_control/requirements.txt
numpy>=1.24.0
scipy>=1.11.0
PyQt5>=5.15.0
matplotlib>=3.7.0
pyyaml>=6.0
